Moved all entities into dedicated package
