+++ src/main/java/Entities/Properties/Restaurant.java
@@ -1,7 +1,7 @@
-package Properties;
+package Entities.Properties;
 
 
-import Users.Owner;
+import Entities.Users.Owner;
 import java.util.ArrayList;
 
 public class Restaurant extends Property {
