+++ src/main/java/Entities/Users/Realtor.java
@@ -1,4 +1,4 @@
-package Users;
+package Entities.Users;
 
 public class Realtor extends User{
     public Realtor(String ID, String name, String password, String contact, String securityQuestion, String securityAnswer) {
