+++ src/main/java/Main.java
@@ -1,10 +1,9 @@
 
 import Exceptions.UndefinedUserType;
 import Managers.PropertyManager;
-import Users.User;
+import Entities.Users.User;
 
 import java.io.IOException;
-import org.json.JSONObject;
 
 public class Main {
     public static void main(String[] args) throws UndefinedUserType, IOException {
